All features added

+++ src/main/java/com/hybridlibrary/services/serviceimpl/BookCopyServiceImpl.java
@@ -1,6 +1,7 @@
 package com.hybridlibrary.services.serviceimpl;
 
 import com.hybridlibrary.dtos.BookCopyDto;
+import com.hybridlibrary.exception.NotFoundException;
 import com.hybridlibrary.models.Book;
 import com.hybridlibrary.models.BookCopy;
 import com.hybridlibrary.models.BookRental;
@@ -14,6 +15,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.convert.ConversionService;
 import org.springframework.stereotype.Service;
+import org.springframework.util.CollectionUtils;
 
 import java.time.LocalDate;
 import java.time.temporal.ChronoUnit;
@@ -41,30 +43,42 @@ public class BookCopyServiceImpl implements BookCopyService {
                 bookCopyRepository.findAll()) {
             bookCopyList.add(conversionService.convert(bookCopy, BookCopyDto.class));
         }
-        log.info("Book copies fetched");
-        return bookCopyList;
+        if (CollectionUtils.isEmpty(bookCopyList)) {
+            throw new NotFoundException("Book copies not found.");
+        } else {
+            log.info("Book copies fetched");
+            return bookCopyList;
+        }
     }
 
     @Override
     public BookCopyDto getOne(Long id) {
-        log.info("Book copy with id {} is listed.", id);
-        return conversionService.convert(bookCopyRepository.getOne(id), BookCopyDto.class);
+        if (bookCopyRepository.existsById(id)) {
+            log.info("Book copy with id {} is listed.", id);
+            return conversionService.convert(bookCopyRepository.getOne(id), BookCopyDto.class);
+        } else {
+            throw new NotFoundException("Book copy with id " + id + " not found.");
+        }
+
     }
 
     @Override
     public List<BookCopyDto> getByBook(Long id) {
         List<BookCopyDto> bookCopyList = new ArrayList<>();
         Book book = bookRepository.getOne(id);
+
         for (BookCopy bookCopy :
                 bookCopyRepository.findByBook(book)) {
             bookCopyList.add(conversionService.convert(bookCopy, BookCopyDto.class));
         }
-        log.info("Copies of the book with id {} are listed.", id);
-        return bookCopyList;
-    }
 
-    public BookCopyDto rentBook(BookCopyDto bookCopyDto) {
-        return null;
+        if (CollectionUtils.isEmpty(bookCopyList)) {
+            throw new NotFoundException("Book copy for book with id " + id + " not found.");
+
+        } else {
+            log.info("Copies of the book with id {} are listed.", id);
+            return bookCopyList;
+        }
     }
 
     @Override
@@ -72,36 +86,37 @@ public class BookCopyServiceImpl implements BookCopyService {
         BookCopy oldBookCopy = bookCopyRepository.getOne(bookCopyDto.getId());
         BookCopy newBookCopy = conversionService.convert(bookCopyDto, BookCopy.class);
 
-        if (newBookCopy.getRented()) {
-            if (oldBookCopy.getRented()) {
-                throw new IllegalArgumentException("You can not rent already rented book.");
-            } else {
-                User user = userRepository.getOne(bookCopyDto.getUserId());
-                oldBookCopy.setUser(user);
-                if (newBookCopy.getRentDate() == null) {
-                    throw new NullPointerException("Rent date can not be null");
+        if (bookCopyRepository.existsById(bookCopyDto.getId())) {
+            if (newBookCopy.getRented()) {
+                if (oldBookCopy.getRented()) {
+                    throw new IllegalArgumentException("You can not rent already rented book.");
                 } else {
-                    oldBookCopy.setRentDate(newBookCopy.getRentDate());
+                    User user = userRepository.getOne(bookCopyDto.getUserId());
+                    oldBookCopy.setUser(user);
+                    oldBookCopy.setRentDate(LocalDate.now());
+                }
+            } else {
+                if (oldBookCopy.getRented()) {
+                    User user = userRepository.getOne(oldBookCopy.getUser().getId());
+                    BookRental bookRental = BookRental.builder()
+                            .user(user)
+                            .book(oldBookCopy.getBook())
+                            .bookCopy(oldBookCopy)
+                            .returnDate(LocalDate.now())
+                            .build();
+                    bookRentalRepository.save(bookRental);
+                    oldBookCopy.setRentDate(null);
+                    oldBookCopy.setUser(null);
                 }
             }
-        } else {
-            if (oldBookCopy.getRented()) {
-                User user = userRepository.getOne(oldBookCopy.getUser().getId());
-                BookRental bookRental = BookRental.builder()
-                        .user(user)
-                        .book(oldBookCopy.getBook())
-                        .bookCopy(oldBookCopy)
-                        .returnDate(LocalDate.now())
-                        .build();
-                bookRentalRepository.save(bookRental);
-                oldBookCopy.setRentDate(null);
-                oldBookCopy.setUser(null);
-            }
-        }
-        oldBookCopy.setRented(newBookCopy.getRented());
+            oldBookCopy.setRented(newBookCopy.getRented());
 
-        log.info("Book copy with id {} is updated.", oldBookCopy.getId());
-        return conversionService.convert(bookCopyRepository.save(oldBookCopy), BookCopyDto.class);
+            log.info("Book copy with id {} is updated.", oldBookCopy.getId());
+            return conversionService.convert(bookCopyRepository.save(oldBookCopy), BookCopyDto.class);
+        } else {
+            throw new NotFoundException("Book copy with id " + bookCopyDto.getId() + " not found.");
+        }
+
     }
 
     @Override
@@ -116,29 +131,29 @@ public class BookCopyServiceImpl implements BookCopyService {
         BookCopy bookCopy = conversionService.convert(bookCopyDto, BookCopy.class);
         Book book = bookRepository.getOne(bookId);
         bookCopy.setBook(book);
-
         if (bookCopy.getRented()) {
             User user = userRepository.getOne(bookCopyDto.getUserId());
             bookCopy.setUser(user);
-           /* if (bookCopy.getRentDate() == null) {
-                throw new NullPointerException("Rent date can not be null");
-            }*/
-           bookCopy.setRentDate(LocalDate.now());
+            bookCopy.setRentDate(LocalDate.now());
         } else {
             bookCopy.setRentDate(null);
         }
-
         return conversionService.convert(bookCopyRepository.save(bookCopy), BookCopyDto.class);
     }
 
-
     @Override
     public BookCopyDto delete(Long id) {
-        log.info("Book copy with id {} is deleted.", id);
-        BookCopy bookCopy = bookCopyRepository.getOne(id);
-        bookRentalRepository.deleteByBookCopy(bookCopy);
-        bookCopyRepository.deleteById(id);
-        return conversionService.convert(bookCopy, BookCopyDto.class);
+
+        if (bookRepository.existsById(id)) {
+            BookCopy bookCopy = bookCopyRepository.getOne(id);
+            bookRentalRepository.deleteByBookCopy(bookCopy);
+            log.info("Book copy with id {} is deleted.", id);
+            bookCopyRepository.deleteById(id);
+            return conversionService.convert(bookCopy, BookCopyDto.class);
+        } else {
+            throw new NotFoundException("Book copy with id " + id + " not found.");
+        }
+
     }
 
     @Override
@@ -148,8 +163,16 @@ public class BookCopyServiceImpl implements BookCopyService {
 
     @Override
     public BookCopyDto findByBookAndId(Long bookId, Long id) {
-        Book book = bookRepository.getOne(bookId);
-        return conversionService.convert(bookCopyRepository.findByBookAndId(book, id), BookCopyDto.class);
+        if (bookCopyRepository.existsById(id) && bookRepository.existsById(bookId)) {
+            Book book = bookRepository.getOne(bookId);
+            return conversionService.convert(bookCopyRepository.findByBookAndId(book, id), BookCopyDto.class);
+        } else {
+            if (bookCopyRepository.existsById(id)) {
+                throw new NotFoundException("Book with id " + bookId + " not found.");
+            } else {
+                throw new NotFoundException("Book copy with id " + id + " not found.");
+            }
+        }
     }
 
     @Override
@@ -159,7 +182,6 @@ public class BookCopyServiceImpl implements BookCopyService {
         for (BookCopy bookCopy : bookCopyRepository.findAll()) {
             if (bookCopy.getRented()) {
                 Long duration = ChronoUnit.DAYS.between(bookCopy.getRentDate(), LocalDate.now());
-
                 if (duration > bookCopy.getBook().getRentPeriod()) {
                     overdueBooks.add(bookCopy);
                 }
@@ -169,9 +191,10 @@ public class BookCopyServiceImpl implements BookCopyService {
         for (BookCopy bookCopy : overdueBooks) {
             bookCopyList.add(conversionService.convert(bookCopy, BookCopyDto.class));
         }
-
-        return bookCopyList;
+        if (CollectionUtils.isEmpty(bookCopyList)) {
+            throw new NotFoundException("There is no overdue book returns.");
+        } else {
+            return bookCopyList;
+        }
     }
-
-
 }
