+++ src/main/java/com/hybridlibrary/services/serviceimpl/UserServiceImpl.java
@@ -1,6 +1,7 @@
 package com.hybridlibrary.services.serviceimpl;
 
 import com.hybridlibrary.dtos.UserDto;
+import com.hybridlibrary.exception.NotFoundException;
 import com.hybridlibrary.models.User;
 import com.hybridlibrary.repositories.UserRepository;
 import com.hybridlibrary.services.UserService;
@@ -8,7 +9,10 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.convert.ConversionService;
 import org.springframework.stereotype.Service;
+import org.springframework.util.CollectionUtils;
 
+import javax.transaction.Transactional;
+import java.beans.Transient;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -28,14 +32,22 @@ public class UserServiceImpl implements UserService {
                 userRepository.findAll()) {
             userDtoList.add(conversionService.convert(user, UserDto.class));
         }
-        log.info("Users fetched");
-        return userDtoList;
+        if (CollectionUtils.isEmpty(userDtoList)) {
+            throw new NotFoundException("There is no any user.");
+        } else {
+            log.info("Users fetched");
+            return userDtoList;
+        }
     }
 
     @Override
     public UserDto getOne(Long id) {
-        log.info("User with id {} is listed", id);
-        return conversionService.convert(userRepository.getOne(id), UserDto.class);
+        if (userRepository.existsById(id)) {
+            log.info("User with id {} is listed", id);
+            return conversionService.convert(userRepository.getOne(id), UserDto.class);
+        } else {
+            throw new NotFoundException("User with id " + id + " not found");
+        }
     }
 
     @Override
@@ -45,15 +57,25 @@ public class UserServiceImpl implements UserService {
                 userRepository.findByUsernameContainingIgnoreCase(username)) {
             userDtoList.add(conversionService.convert(user, UserDto.class));
         }
-        log.info("Books which title contains '{}' are listed", username);
-        return userDtoList;
+        if (CollectionUtils.isEmpty(userDtoList)) {
+            throw new NotFoundException("There is no any user which username contains" + username);
+        } else {
+            log.info("Users which username contains '{}' are listed", username);
+            return userDtoList;
+        }
+
     }
 
     @Override
     public UserDto update(UserDto userDto) {
         User user = conversionService.convert(userDto, User.class);
-        log.info("{} is updated.", user);
-        return conversionService.convert(userRepository.save(user), UserDto.class);
+        if (userRepository.existsById(userDto.getId())) {
+            log.info("{} is updated.", user);
+            return conversionService.convert(userRepository.save(user), UserDto.class);
+        } else {
+            throw new NotFoundException("There is no any user with id " + user.getId());
+        }
+
     }
 
     @Override
@@ -66,9 +88,14 @@ public class UserServiceImpl implements UserService {
     @Override
     public UserDto delete(Long id) {
         User user = userRepository.getOne(id);
-        log.info("User with id {} is deleted.", id);
-        userRepository.deleteById(id);
-        return conversionService.convert(user, UserDto.class);
+        if (userRepository.existsById(id)) {
+            log.info("User with id {} is deleted.", id);
+            userRepository.deleteById(id);
+            return conversionService.convert(user, UserDto.class);
+        } else {
+            throw new NotFoundException("User with id " + id + " not found.");
+        }
+
     }
 
     @Override
@@ -78,9 +105,13 @@ public class UserServiceImpl implements UserService {
 
     public UserDto createPassword(Long id, String password) {
         User user = userRepository.getOne(id);
-        user.setPassword(password);
-        update(conversionService.convert(user, UserDto.class));
-        return conversionService.convert(user, UserDto.class);
+        if (userRepository.existsById(id)) {
+            user.setPassword(password);
+            UserDto userDto = conversionService.convert(user, UserDto.class);
+            userRepository.executeUpdate(password, id);
+            return userDto;
+        } else {
+            throw new NotFoundException("User with id " + id + " not found.");
+        }
     }
-
 }
